+++ src/promptflow/tests/sdk_cli_test/unittests/test_config.py
@@ -1,9 +1,15 @@
 # ---------------------------------------------------------
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # ---------------------------------------------------------
-import pytest
+from pathlib import Path
 
-from promptflow._sdk._configuration import Configuration
+import pytest
+from azure.ai.ml.constants._common import AZUREML_RESOURCE_PROVIDER, RESOURCE_ID_FORMAT
+
+from promptflow._sdk._configuration import ConfigFileNotFound, Configuration
+from promptflow._utils.context_utils import _change_working_dir
+
+CONFIG_DATA_ROOT = Path(__file__).parent.parent.parent / "test_configs" / "configs"
 
 
 @pytest.fixture
@@ -14,21 +20,14 @@ def config():
 @pytest.mark.unittest
 class TestConfig:
     def test_set_config(self, config):
-        config._set_config("a.b.c.test_key", "test_value")
-        assert config._get_config("a.b.c.test_key") == "test_value"
-        assert config.config == {"a": {"b": {"c": {"test_key": "test_value"}}}}
+        config.set_config("a.b.c.test_key", "test_value")
+        assert config.get_config("a.b.c.test_key") == "test_value"
+        # global config may contain other keys
+        assert config.config["a"] == {"b": {"c": {"test_key": "test_value"}}}
 
     def test_get_config(self, config):
-        config._set_config("test_key", "test_value")
-        assert config._get_config("test_key") == "test_value"
-
-    def test_get_telemetry_consent(self, config):
-        config.set_telemetry_consent(True)
-        assert config.get_telemetry_consent() is True
-
-    def test_set_telemetry_consent(self, config):
-        config.set_telemetry_consent(True)
-        assert config.get_telemetry_consent() is True
+        config.set_config("test_key", "test_value")
+        assert config.get_config("test_key") == "test_value"
 
     def test_get_or_set_installation_id(self, config):
         user_id = config.get_or_set_installation_id()
@@ -37,3 +36,23 @@ class TestConfig:
     def test_config_instance(self, config):
         new_config = Configuration.get_instance()
         assert new_config is config
+
+    def test_get_workspace_from_config(self):
+        # New instance instead of get_instance() to avoid side effect
+        conf = Configuration()
+        conf.set_config("connection.provider", "azureml")
+        # Test config within flow folder
+        target_folder = CONFIG_DATA_ROOT / "mock_flow1"
+        with _change_working_dir(target_folder):
+            config1 = conf.get_connection_provider()
+        assert config1 == "azureml:" + RESOURCE_ID_FORMAT.format("sub1", "rg1", AZUREML_RESOURCE_PROVIDER, "ws1")
+        # Test config using flow parent folder
+        target_folder = CONFIG_DATA_ROOT / "mock_flow2"
+        with _change_working_dir(target_folder):
+            config2 = conf.get_connection_provider()
+        assert config2 == "azureml:" + RESOURCE_ID_FORMAT.format(
+            "sub_default", "rg_default", AZUREML_RESOURCE_PROVIDER, "ws_default"
+        )
+        # Test config not found
+        with pytest.raises(ConfigFileNotFound):
+            Configuration._get_workspace_from_config(path=CONFIG_DATA_ROOT.parent)